optimize chunkToPixels colour blending
Modify chunkToPixels to use "front to back" alpha blending rather than
"back to front" blending.
With the back to front blending used previously the first opaque block
in the column needed to be found first so that the algorithm could blend
upwards from that point.
With front to back blending we can simply start at the topmost block and
blend until an opaque block is reached, so it saves a search through the
block column.