Add merge to tiled image set
Now when areas larger than 8192x8192 are selected to be merged, a set of
images are produced.

Tried adding a BMP image implementation that was meant to allow writing
very large images, but it is slow and produces huge uncompressed files.

Also added a new class MapViewRequest that is used to save the important
parts of a MapView instance for use in requesting regions to be loaded
and checking what view is already loaded.