Stop using ModLoader compatibility code
Also experimenting with alternatives to using Anvil region files to save
the map. It's not very efficient if all we need is the top layer of the
world.

MapChunk was meant to be a way to efficiently store only the blocks we
need from a chunk using Run Length Encoding. The idea being that RLE is
fast enough that the compressed copy can be stored in memory rather than
a decompressed 256x16x16 block array for each chunk.

Will probably delete MapChunk next commit but I wanted it to be in the
Git history for possible future reference.

The 1.7 update is going to be painful with the dynamic block ID's, but
is probably a good point to change the save format.